test: rearrange unit tests to one test file per component
