+++ tests/IgniteUI.Blazor.Tests/DropdownTests.cs
@@ -115,6 +115,15 @@ public class DropdownTests : ComponentWithContractTestBase<IgbDropdown>
     }
 
     [Fact]
+    public void Dropdown_ChildContent_Renders()
+    {
+        var cut = Render<IgbDropdown>(parameters =>
+            parameters.AddChildContent("<igc-dropdown-item>Item</igc-dropdown-item>"));
+
+        Assert.Contains("Item", cut.Find("igc-dropdown").InnerHtml);
+    }
+
+    [Fact]
     public void Dropdown_InheritsFromBaseRendererControl()
     {
         Assert.True(typeof(IgbDropdown).IsSubclassOf(typeof(BaseRendererControl)));
@@ -168,11 +177,59 @@ public class DropdownItemTests : BlazorComponentTestBase
     }
 
     [Fact]
+    public void DropdownItem_Active_RendersAttribute()
+    {
+        var cut = Render<IgbDropdownItem>(parameters =>
+            parameters.Add(p => p.Active, true));
+
+        var element = cut.Find("igc-dropdown-item");
+        Assert.NotNull(element.GetAttribute("active"));
+    }
+
+    [Fact]
     public void DropdownItem_ChildContent_Renders()
     {
         var cut = Render<IgbDropdownItem>(parameters =>
             parameters.AddChildContent("Option 1"));
 
         Assert.Contains("Option 1", cut.Markup);
+    }
+}
+
+public class DropdownGroupTests : BlazorComponentTestBase
+{
+    [Fact]
+    public void DropdownGroup_RendersCorrectElement()
+    {
+        var cut = Render<IgbDropdownGroup>();
+        cut.Find("igc-dropdown-group").Should_Exist();
+    }
+
+    [Fact]
+    public void DropdownGroup_ChildContent_Renders()
+    {
+        var cut = Render<IgbDropdownGroup>(parameters =>
+            parameters.AddChildContent("<igc-dropdown-item>Grouped</igc-dropdown-item>"));
+
+        Assert.Contains("Grouped", cut.Find("igc-dropdown-group").InnerHtml);
+    }
+}
+
+public class DropdownHeaderTests : BlazorComponentTestBase
+{
+    [Fact]
+    public void DropdownHeader_RendersCorrectElement()
+    {
+        var cut = Render<IgbDropdownHeader>();
+        cut.Find("igc-dropdown-header").Should_Exist();
+    }
+
+    [Fact]
+    public void DropdownHeader_ChildContent_Renders()
+    {
+        var cut = Render<IgbDropdownHeader>(parameters =>
+            parameters.AddChildContent("Category"));
+
+        Assert.Contains("Category", cut.Find("igc-dropdown-header").InnerHtml);
     }
 }
