+++ tests/IgniteUI.Blazor.Tests/LinearProgressTests.cs
@@ -94,57 +94,14 @@ public class LinearProgressTests : BlazorComponentTestBase
         var element = cut.Find("igc-linear-progress");
         Assert.Equal("{0}%", element.GetAttribute("label-format"));
     }
-}
-
-public class CircularProgressTests : BlazorComponentTestBase
-{
-    [Fact]
-    public void CircularProgress_RendersCorrectElement()
-    {
-        var cut = Render<IgbCircularProgress>();
-        Assert.NotNull(cut.Find("igc-circular-progress"));
-    }
 
     [Fact]
-    public void CircularProgress_TypeMetadata_IsCorrect()
+    public void LinearProgress_AnimationDuration_RendersAttribute()
     {
-        var progress = new IgbCircularProgress();
-        Assert.Equal("WebCircularProgress", progress.Type);
-    }
+        var cut = Render<IgbLinearProgress>(parameters =>
+            parameters.Add(p => p.AnimationDuration, 500));
 
-    [Fact]
-    public void CircularProgress_InheritsFromProgressBase()
-    {
-        Assert.True(typeof(IgbCircularProgress).IsSubclassOf(typeof(IgbProgressBase)));
-    }
-
-    [Fact]
-    public void CircularProgress_Value_RendersAttribute()
-    {
-        var cut = Render<IgbCircularProgress>(parameters =>
-            parameters.Add(p => p.Value, 60.0));
-
-        var element = cut.Find("igc-circular-progress");
-        Assert.Equal("60", element.GetAttribute("value"));
-    }
-
-    [Fact]
-    public void CircularProgress_Indeterminate_RendersAttribute()
-    {
-        var cut = Render<IgbCircularProgress>(parameters =>
-            parameters.Add(p => p.Indeterminate, true));
-
-        var element = cut.Find("igc-circular-progress");
-        Assert.NotNull(element.GetAttribute("indeterminate"));
-    }
-
-    [Fact]
-    public void CircularProgress_Max_RendersAttribute()
-    {
-        var cut = Render<IgbCircularProgress>(parameters =>
-            parameters.Add(p => p.Max, 150.0));
-
-        var element = cut.Find("igc-circular-progress");
-        Assert.Equal("150", element.GetAttribute("max"));
+        var element = cut.Find("igc-linear-progress");
+        Assert.Equal("500", element.GetAttribute("animation-duration"));
     }
 }
