+++ tests/IgniteUI.Blazor.Tests/SliderTests.cs
@@ -137,4 +137,103 @@ public class SliderTests : ComponentWithContractTestBase<IgbSlider>
         var element = cut.Find("igc-slider");
         Assert.Equal("80", element.GetAttribute("upper-bound"));
     }
+
+    [Fact]
+    public void Slider_PrimaryTicks_RendersAttribute()
+    {
+        var cut = Render<IgbSlider>(parameters =>
+            parameters.Add(p => p.PrimaryTicks, 5));
+
+        var element = cut.Find("igc-slider");
+        Assert.Equal("5", element.GetAttribute("primary-ticks"));
+    }
+
+    [Fact]
+    public void Slider_SecondaryTicks_RendersAttribute()
+    {
+        var cut = Render<IgbSlider>(parameters =>
+            parameters.Add(p => p.SecondaryTicks, 3));
+
+        var element = cut.Find("igc-slider");
+        Assert.Equal("3", element.GetAttribute("secondary-ticks"));
+    }
+
+    [Fact]
+    public void Slider_HidePrimaryLabels_RendersAttribute()
+    {
+        var cut = Render<IgbSlider>(parameters =>
+            parameters.Add(p => p.HidePrimaryLabels, true));
+
+        var element = cut.Find("igc-slider");
+        Assert.NotNull(element.GetAttribute("hide-primary-labels"));
+    }
+
+    [Fact]
+    public void Slider_HideSecondaryLabels_RendersAttribute()
+    {
+        var cut = Render<IgbSlider>(parameters =>
+            parameters.Add(p => p.HideSecondaryLabels, true));
+
+        var element = cut.Find("igc-slider");
+        Assert.NotNull(element.GetAttribute("hide-secondary-labels"));
+    }
+
+    [Fact]
+    public void Slider_TickOrientation_Mirror()
+    {
+        var cut = Render<IgbSlider>(parameters =>
+            parameters.Add(p => p.TickOrientation, SliderTickOrientation.Mirror));
+
+        var element = cut.Find("igc-slider");
+        Assert.Equal("mirror", element.GetAttribute("tick-orientation"));
+    }
+
+    [Fact]
+    public void Slider_Locale_RendersAttribute()
+    {
+        var cut = Render<IgbSlider>(parameters =>
+            parameters.Add(p => p.Locale, "en-US"));
+
+        var element = cut.Find("igc-slider");
+        Assert.Equal("en-US", element.GetAttribute("locale"));
+    }
+
+    [Fact]
+    public void Slider_ValueFormat_RendersAttribute()
+    {
+        var cut = Render<IgbSlider>(parameters =>
+            parameters.Add(p => p.ValueFormat, "{0}%"));
+
+        var element = cut.Find("igc-slider");
+        Assert.Equal("{0}%", element.GetAttribute("value-format"));
+    }
+
+    [Fact]
+    public void Slider_Invalid_RendersAttribute()
+    {
+        var cut = Render<IgbSlider>(parameters =>
+            parameters.Add(p => p.Invalid, true));
+
+        var element = cut.Find("igc-slider");
+        Assert.NotNull(element.GetAttribute("invalid"));
+    }
+}
+
+public class SliderLabelTests : BlazorComponentTestBase
+{
+    [Fact]
+    public void SliderLabel_RendersCorrectElement()
+    {
+        var cut = Render<IgbSliderLabel>();
+        cut.Find("igc-slider-label").Should_Exist();
+    }
+
+    [Fact]
+    public void SliderLabel_ChildContent_Renders()
+    {
+        var cut = Render<IgbSliderLabel>(parameters =>
+            parameters.AddChildContent("Low"));
+
+        Assert.Contains("Low", cut.Find("igc-slider-label").InnerHtml);
+    }
 }
